Mems sensor filtering functional
